reduce noise in entry point
