+++ lib/levels.js
@@ -1,8 +1,9 @@
 'use strict'
+const flatstr = require('flatstr')
 
-var flatstr = require('flatstr')
+const { noop, genLog } = require('./tools')
 
-var levels = {
+const levels = {
   fatal: 60,
   error: 50,
   warn: 40,
@@ -11,21 +12,28 @@ var levels = {
   trace: 10
 }
 
-var nums = Object.keys(levels).reduce(function (o, k) {
+const levelMethods = {
+  fatal: genLog(levels.fatal),
+  error: genLog(levels.error),
+  warn: genLog(levels.warn),
+  info: genLog(levels.info),
+  debug: genLog(levels.debug),
+  trace: genLog(levels.trace)
+}
+
+const nums = Object.keys(levels).reduce(function (o, k) {
   o[levels[k]] = k
   return o
 }, {})
 
 // level string cache
-var lscache = Object.keys(nums).reduce(function (o, k) {
+const lscache = Object.keys(nums).reduce(function (o, k) {
   o[k] = flatstr('{"level":' + Number(k))
   return o
 }, {})
 
 function isStandardLevel (level) {
-  if (level === Infinity) {
-    return true
-  }
+  if (level === Infinity) return true
   switch (level) {
     case 'fatal':
     case 'error':
@@ -54,10 +62,70 @@ function isStandardLevelVal (val) {
   }
 }
 
+function getLevelVal () {
+  return this._levelVal
+}
+function setLevelVal (num) {
+  if (typeof num === 'string') { return this._setLevel(num) }
+
+  if (this.emit) {
+    this.emit('level-change', this.levels.labels[num], num, this.levels.labels[this._levelVal], this._levelVal)
+  }
+
+  this._levelVal = num
+
+  for (var key in this.levels.values) {
+    if (num > this.levels.values[key]) {
+      this[key] = noop
+      continue
+    }
+    this[key] = isStandardLevel(key) ? levelMethods[key] : genLog(this.levels.values[key])
+  }
+}
+
+function _setLevel (level) {
+  if (typeof level === 'number') {
+    if (!isFinite(level)) {
+      throw Error('unknown level ' + level)
+    }
+    level = this.levels.labels[level]
+  }
+  if (!this.levels.values[level]) {
+    throw Error('unknown level ' + level)
+  }
+  this.levelVal = this.levels.values[level]
+}
+
+function _getLevel (level) {
+  return this.levels.labels[this.levelVal]
+}
+
+function addLevel (name, lvl) {
+  if (this.levels.values.hasOwnProperty(name)) return false
+  if (this.levels.labels.hasOwnProperty(lvl)) return false
+  this.levels.values[name] = lvl
+  this.levels.labels[lvl] = name
+  this._lscache[lvl] = flatstr('{"level":' + Number(lvl))
+  this[name] = lvl < this._levelVal ? noop : genLog(lvl)
+  return true
+}
+
+function isLevelEnabled (logLevel) {
+  const logLevelVal = this.levels.values[logLevel]
+  return logLevelVal && (logLevelVal >= this._levelVal)
+}
+
 module.exports = {
-  levels: levels,
-  nums: nums,
-  lscache: lscache,
-  isStandardLevel: isStandardLevel,
-  isStandardLevelVal: isStandardLevelVal
+  isStandardLevelVal,
+  isStandardLevel,
+  lscache,
+  levels,
+  nums,
+  levelMethods,
+  setLevelVal,
+  getLevelVal,
+  _getLevel,
+  _setLevel,
+  addLevel,
+  isLevelEnabled
 }
