+++ lib/tools.js
@@ -2,29 +2,13 @@
 
 var format = require('quick-format-unescaped')
 var util = require('util')
-var levels = require('./levels')
 var serializers = require('pino-std-serializers')
-var isStandardLevelVal = levels.isStandardLevelVal
-var isStandardLevel = levels.isStandardLevel
 
 function noop () {}
 
 function copy (a, b) {
   for (var k in b) { a[k] = b[k] }
   return a
-}
-
-function applyOptions (self, opts) {
-  self.serializers = opts.serializers
-  self.chindings = opts.chindings
-
-  if (opts.level && opts.levelVal) {
-    var levelIsStandard = isStandardLevel(opts.level)
-    var valIsStandard = isStandardLevelVal(opts.levelVal)
-    if (valIsStandard) throw Error('level value is already used: ' + opts.levelVal)
-    if (levelIsStandard === false && valIsStandard === false) self.addLevel(opts.level, opts.levelVal)
-  }
-  self._setLevel(opts.level)
 }
 
 function asMetaWrapper (pretty, dest) {
@@ -63,7 +47,7 @@ function asMetaWrapper (pretty, dest) {
   }
 }
 
-function defineLevelsProperty (onObject) {
+function defineLevelsProperty (onObject, levels) {
   Object.defineProperty(onObject, 'levels', {
     value: {
       values: copy({}, levels.levels),
@@ -124,11 +108,107 @@ function genLog (z) {
   }
 }
 
+function getPrettyStream (opts, prettifier, dest) {
+  if (prettifier && typeof prettifier === 'function') {
+    return asMetaWrapper(prettifier(opts), dest)
+  }
+  try {
+    var prettyFactory = require('pino-pretty')
+    return asMetaWrapper(prettyFactory(opts), dest)
+  } catch (e) {
+    throw Error('Missing `pino-pretty` module: `pino-pretty` must be installed separately')
+  }
+}
+
+// magically escape strings for json
+// relying on their charCodeAt
+// everything below 32 needs JSON.stringify()
+// 34 and 92 happens all the time, so we
+// have a fast case for them
+function asString (str) {
+  var result = ''
+  var last = 0
+  var l = str.length
+  var point = 255
+  if (l > 42) {
+    return JSON.stringify(str)
+  }
+  for (var i = 0; i < l && point >= 32; i++) {
+    point = str.charCodeAt(i)
+    if (point === 34 || point === 92) {
+      result += str.slice(last, i) + '\\'
+      last = i
+    }
+  }
+  if (last === 0) {
+    result = str
+  } else {
+    result += str.slice(last)
+  }
+  return point < 32 ? JSON.stringify(str) : '"' + result + '"'
+}
+
+function asJson (obj, msg, num, time) {
+  // to catch both null and undefined
+  var hasObj = obj !== undefined && obj !== null
+  var objError = hasObj && obj instanceof Error
+  msg = !msg && objError ? obj.message : msg || undefined
+  var data = this._lscache[num] + time
+  if (msg !== undefined) {
+    data += this.messageKeyString + asString('' + msg)
+  }
+  // we need the child bindings added to the output first so that logged
+  // objects can take precedence when JSON.parse-ing the resulting log line
+  data = data + this.chindings
+  var value
+  if (hasObj) {
+    var notHasOwnProperty = obj.hasOwnProperty === undefined
+    if (objError) {
+      data += ',"type":"Error","stack":' + this.stringify(obj.stack)
+    }
+    // if global serializer is set, call it first
+    if (this.serializers[Symbol.for('pino.*')]) {
+      obj = this.serializers[Symbol.for('pino.*')](obj)
+    }
+    for (var key in obj) {
+      value = obj[key]
+      if ((notHasOwnProperty || obj.hasOwnProperty(key)) && value !== undefined) {
+        value = (this.stringifiers[key] || this.stringify)(this.serializers[key] ? this.serializers[key](value) : value)
+        if (value !== undefined) {
+          data += ',"' + key + '":' + value
+        }
+      }
+    }
+  }
+  return data + this.end
+}
+
+function asChindings (that, bindings) {
+  if (!bindings) {
+    throw Error('missing bindings for child Pino')
+  }
+  var key
+  var value
+  var data = that.chindings
+  if (that.serializers[Symbol.for('pino.*')]) {
+    bindings = that.serializers[Symbol.for('pino.*')](bindings)
+  }
+  for (key in bindings) {
+    value = bindings[key]
+    if (key !== 'level' && key !== 'serializers' && bindings.hasOwnProperty(key) && value !== undefined) {
+      value = that.serializers[key] ? that.serializers[key](value) : value
+      data += ',"' + key + '":' + (that.stringifiers[key] || that.stringify)(value)
+    }
+  }
+  return data
+}
+
 module.exports = {
-  noop: noop,
-  copy: copy,
-  applyOptions: applyOptions,
-  asMetaWrapper: asMetaWrapper,
-  defineLevelsProperty: defineLevelsProperty,
-  genLog: genLog
+  copy,
+  defineLevelsProperty,
+  noop,
+  getPrettyStream,
+  asChindings,
+  asJson,
+  genLog
 }
